update solid ssr to type module
