+++ .changeset/cyan-seas-yawn.md
@@ -1,0 +1,5 @@
+---
+"solid-ssr": minor
+---
+
+update solid ssr to type module
