+++ packages/solid-ssr/examples/async/index.js
@@ -1,5 +1,5 @@
 import express from "express";
-import path from "path";
+import url from "url";
 
 import { renderToStringAsync } from "solid-js/web";
 import App from "../shared/src/components/App";
@@ -7,7 +7,7 @@ import App from "../shared/src/components/App";
 const app = express();
 const port = 8080;
 
-app.use(express.static(path.join(__dirname, "../public")));
+app.use(express.static(url.fileURLToPath(new URL("../public", import.meta.url))));
 
 app.get("*", async (req, res) => {
   let result;
