+++ packages/solid-ssr/examples/ssg/export.js
@@ -1,7 +1,9 @@
-const path = require("path");
-const renderStatic = require("../../static");
+import path from "path";
+import url from "url";
+import renderStatic from "../../static/index.js";
 
 const PAGES = ["index", "profile", "settings"];
+const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
 const pathToServer = path.resolve(__dirname, "lib/index.js");
 const pathToPublic = path.resolve(__dirname, "public");
