+++ packages/solid-ssr/examples/stream/rollup.config.js
@@ -10,7 +10,7 @@ export default [
     output: [
       {
         dir: "examples/stream/lib",
-        format: "cjs"
+        format: "esm"
       }
     ],
     external: ["solid-js", "solid-js/web", "path", "express", "stream"],
