+++ packages/solid-ssr/static/index.js
@@ -1,15 +1,19 @@
-const path = require("path");
-const execFile = require("util").promisify(require("child_process").execFile);
+import { resolve } from "path";
+import { fileURLToPath } from "url";
+import { execFile } from "child_process";
+import { promisify } from "util";
 
-const pathToRunner = path.resolve(__dirname, "writeToDisk.js");
+const exec = promisify(execFile);
+const __dirname = fileURLToPath(new URL(".", import.meta.url));
+const pathToRunner = resolve(__dirname, "writeToDisk.js");
 
 async function run({ entry, output, url }) {
-  const { stdout, stderr } = await execFile("node", [pathToRunner, entry, output, url, "--trace-warnings"]);
+  const { stdout, stderr } = await exec("node", [pathToRunner, entry, output, url, "--trace-warnings"]);
   if (stdout.length) console.log(stdout);
   if (stderr.length) console.log(stderr);
 }
 
-module.exports = async function renderStatic(config) {
+export default async function renderStatic(config) {
   if (Array.isArray(config)) {
     await Promise.all(config.map(run));
   } else await run(config);
